+++ src/main/java/ua/foxminded/kucherenko/task3/repositories/StudentRepository.java
@@ -1,17 +1,14 @@
 package ua.foxminded.kucherenko.task3.repositories;
 
-import jakarta.transaction.Transactional;
 import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.data.jpa.repository.Modifying;
 import org.springframework.data.jpa.repository.Query;
 import org.springframework.stereotype.Repository;
-import ua.foxminded.kucherenko.task3.models.Role;
 import ua.foxminded.kucherenko.task3.models.Student;
 
 import java.util.List;
 
 @Repository
-public interface StudentRepository extends JpaRepository<Student, Integer> {
+public interface StudentRepository extends JpaRepository<Student, Long> {
     @Query("""
             SELECT NEW Student(s.id, s.firstName, s.lastName)
             FROM Student s
